feat: add style prop support and centralize Hero interface
- Add React.CSSProperties style prop to Heading and Paragraph components
- Move HeroSectionProps interface to centralized types file for consistency
- Fixes TypeScript errors when using textShadow in Hero component
